Reset network details when lookupNetwork request fails (#1403)
`lookupNetwork`, a method in NetworkController, collects information
about the network via two requests and stores that information in state.
If either of these requests fails, it means that this step was
unsuccessful, and we should reset all of the pieces of state that would
ordinarily get updated after a call to `lookupNetwork` in order to match
the initial state of the controller. The network ID and network status
are correctly reset, but the network details aren't. To be specific,
`networkDetails.EIPS[1559]` will be set to false to indicate that the
network doesn't support EIP-1559. This isn't completely accurate, as we
actually don't know that. This commit fixes this problem by removing the
property instead of setting it to false.