Improve test for NC getEthQuery action
The "happy path" test for NetworkController's `getEthQuery` action
verifies the behavior of the action by mocking the `eth-query` module
to return a fake EthQuery instance and asserting that the action returns
this instance. While this is very straightforward, it ends up
interfering with the provider initialization routine when the test is
set up (as NetworkController uses `eth-query` under the hood to talk to
the network, and now it's not doing anything). For some reason, this has
no negative consequences here on the core side, but when this test is
copied to the extension side, it fails to run for that reason.

Instead of mocking `eth-query`, we simply attempt to use the EthQuery
instance that the `getEthQuery` action returns in order to talk to the
network and verify that the result is the same result we use the mock
the network itself.