Add missing tests for getProviderAndBlockTracker (#1409)
These tests exists within the test suite for the extension version of
NetworkController, but not here.

Note that in this case, if no provider has been initialized, the
provider and block tracker are undefined instead of null. This is due to
slight differences in the constructor between this version and
extension.