Add loadBackup method to NetworkController
The extension offers a way to back up user customized settings in the
wallet and then restoring those settings later from a previous backup.
One of the settings that is preserved in this backup is the list of
custom networks. When restoring this list, the extension attempts to
update NetworkController's state directly. This worked in the extension
version of NetworkController, but it won't work with the version here in
core, as the `update` method in a controller that inherits from
BaseControllerV2 cannot be called from outside the controller itself.

To fix this, this commit adds a new method to NetworkController,
`loadBackup`, which ports the same functionality from the extension.