Encourage better PR descriptions (#1425)
This repo contains a PR template to help contributors understand how to
fill out the description for a new PR. The template is intended more for
reference, however, and so it does not go into detail on the differences
between the Description and Changes sections. This leads to PR
descriptions that lack the detail that the Shared Libraries team would
prefer in order to ensure that pull requests can be reviewed more
thoroughly and that changelogs can be compiled more quickly and contain
more high-quality information after a release is issued.

This commit attempts to improve pull request descriptions going forward
by adding a new section to the Contributing document. It also updates
the pull request template to:

- Rename Description to Explanation so as not to overload the term
  "description"
- Rename Changes to Changelog so as not to overload the term "changes"
- Move up References, as it's most commonly provided after the
  explanation
- Update both Explanation and Changelog to include the same details
  mentioned in Contributing
- Update Changelog to ensure that changes are divided by package name