Simplify network types accepted by TransactionController
According to its types, TransactionController takes the proxy versions
of the provider and block tracker that NetworkController exposes.
However, this isn't really necessary, and in fact this introduces churn,
because if anything around that proxy changes on the NetworkController
side, then the tests for TransactionController need to change too.
By relaxing the types so that TransactionController only need accept the
_actual_ provider and block tracker interfaces rather than the proxies,
it prevents the tests from having to be unnecessary updated and makes
things simpler for consumers as well.

To make this change, `@metamask/network-controller` needed to be updated
to not only expose its Provider type but also its BlockTracker type.