fix(controller-utils): Fix query function
The `query` function was updated in #1266 to use `hasProperty` has part
of an effect to improve the types being used. Unfortunately this had
the unexpected affected of breaking this condition because
`hasProperty` doesn't look up the prototype chain, but the method we
were checking for was on the prototype. `hasProperty` is only meant to
be used with plain objects.

The condition has been updated to use `in`, and the tests have been
updated to use a more realistic EthQuery mock that has methods as
prototypes.

To fix various type errors, an old inlined EthQueryLike type has been
replaced by the real EthQuery type. This required adding `eth-query` as
a dependency to the controller utils package.