Improve error message when cache fails during release
A cache miss during the release workflow guarantees that the release
will fail. The cache has the build; without it, we have nothing to
release.

The release workflows have been updated to fail on cache miss during
the release workflow, ensuring that we spend less time diagnosing what
failed in this scenario. We can proceed directly to retrying to
workflow if this happens.