Prevent event publish from throwing error (#1475)
BaseControllerV2 subscribers can no longer interrupt the flow of
execution when they throw an error. Instead, any errors from a
subscriber are thrown in a timeout hander. This ensures that the errors
are still logged to the console and captured by Sentry without
breaking the controller the event was published from.

Event subscribers should never throw. Any case where it does throw is a
bug that may leave the wallet in an unrecoverable state. Originally we
decided to not prevent event publishing from throwing because we want
to be certain this never happens, and letting the error bubble up
seemed like the "noisiest" way of handling it.

However, we've found that the current strategy does inadvertantely make
these errors hard to discover sometimes. If the method that triggered
the event to be published has its errors captured or suppressed
already, these errors can be difficult to discover. Adding additional
logging is critical for helping us discover these.

The error no longer bubbles up because that was causing state
corruption to spread between controllers. For example, we saw a case
where it appears as through a network controller subscriber threw an
error and caused the network controller to enter an invalid state. It
seems like a good idea to prevent corruption from spreading, so long as
we still have adequate visibility for the error.