Prevent bug where incorrect symbol used for token rates
There is a bug present in the TokenRatesController where a network
switch can result in token rates being requested for the wrong network.

This can happen when the network state update event comes before
currency rate controller state update event. The controller only works
correctly if the currency rate controller state update event happens
first.

Instead of getting the ticker from the currency rate controller, we now
get it directly from the network controller. This simplifies the
controller and ensures that we always use the currency symbol
corresponding with the current chain ID.