Fix initial native currency `TokenRatesController` configuration
The `TokenRatesController` was always initialized with `eth` as the
default `nativeCurrency` configuration. This could result in the wrong
rates being retrieved if the current network has a different currency
symbol.

The constructor has been updated to require a `ticker` parameter, used
to initialize the `nativeCurrency` configuation for the current
network at time of construction.