Extend set of allowed release commit messages (#1494)
When creating a new release branch via `create-release-branch`, it is
common for an engineer to add a new commit to the branch to adjust the
changelogs and then push up the branch. Due to how GitHub works and how
we've set up this repo, this results in creating a PR with the title
"Release/X.Y.Z", and that will be the final commit message when the PR
is squashed & merged. Unfortunately a commit called "Release/X.Y.Z" is
not considered a release commit by our `is-release` release action, as
we've configured it to recognize "Release X.Y.Z" (note the lack of
slash), and so this will not result in a new release being published.

To address this, this commit adds "Release/X.Y.Z" as a valid release
commit message.