Make provider config property `ticker` mandatory (#1495)
The `ticker` property of the provider configuration is mandatory for
custom network configurations, and always set for built-in networks.
The type has been updated to reflect that it should always be set. This
ensures that we don't need to handle the "missing ticker" case
throughout the codebase.

The initial state was missing the `ticker` property, so it's possible
that it will be missing for users. It's possible users have a missing
`ticker` property on `providerConfig` if they added a custom network
prior to us making this mandatory, and haven't switched since. As a
precaution we should set `providerConfig.ticker` to `ETH` in a
migration if it is missing. This should preserve the current behavior
(which is to assume the currency is ETH if it's missing).