Remove unnecessary `babel-runtime` dependencies
We included `babel-runtime` as a dependency to a few packages because
we thought it was an undocumented dependency of `eth-query`, but we had
gotten it mixed up with `ethjs-query`. Really we only need this as a
dependency in the transaction controller.

The constraints have been updated, and all unnecessary references to
`babel-runtime` have been removed.