Fix another `TokenRatesController` event ordering bug (#1511)
* Fix another `TokenRatesController` event ordering bug

When the network switches, the `TokenRatesController` was relying upon
the combination of two state updates (from the network controller and
the tokens controller), both of which were needed in order to ensure
the controller configuration was valid. These events don't arrive at
the same time, meaning that the controller was potentially making one
invalid call each network switch.

Additionally, there is a risk that the second event could result in no
update (e.g. because the destination network isn't supported, or
because there were no tokens), resulting in the data from the invalid
call remaining in state.

This has been prevented by updating the controller to use the
`allTokens` and `allDetectedTokens` state rather than the `tokens`
state. A network switch now triggers a single event, and the controller
configuration never enters an invalid state.

* Prevent update while polling inactive, and add tests