Improve error handling when state metadata is missing (#1529)
* Improve error handling when state metadata is missing

The state derivation functions will throw an error today if any state
properties are missing metadata. This interrupts the state operation
completely. There are a few state operations that we want to ensure are
never interrupted, even if some data is invalid, so this behavior has
been changed.

Instead we will skip any state properties with missing metadata, and
throw an error in a timeout handler so that the problem is still
captured.

* Add missing sinon restore steps