Extract `mockNetwork` from tests (#1508)
Mocking JSON-RPC requests in tests is somewhat inconvenient due to
needing to spell out the full path for an Infura endpoint and specify
full JSON-RPC request and response objects, but to make this easier,
we've written a small wrapper around Nock. Currently we use this in one
test within the `network-controller` package, but we plan to use this in
a different package, so this commit extracts the class responsible for
doing this. This also makes a small improvement to the API and adds a
function that makes the class even easier to use.