Mock all network requests (#1499)
By default HTTP requests are allowed to be made in tests. Because
requests are not guaranteed to succeed, this can cause intermittently
failing CI builds, forcing engineers to wait until PRs are ready to
merge.

This commit addresses this problem by enabling Nock globally and then
modifying each test that hits the network to create proper mocks. This
was somewhat difficult to do with the assets controller tests as they
involve interactions with smart contracts on Mainnet.

Also, Nock was upgraded to 13.3.x, as without this change,
`nock.recorder`, used for debugging, was not working correctly.