eth-query ^2.1.2 -> @metamask/eth-query ^3.0.1 (#1546)
The `eth-query` package does not ship with TypeScript type definitions,
and none have been published under `@types/eth-query`. This results in
TypeScript failing to build packages in this repo that want to use
`eth-query`. We've attempted to mitigate this issue by adding our own
type definitions for `eth-query` to this repo, but since they aren't
included in the published version of the packages in question, it means
that any other project that wants to use those packages must supply
their own type definitions for `eth-query`.

We have republished `eth-query` as `@metamask/eth-query`, which _does_
ship with type definitions. This removes the need for this repo or
any consumers of these packages to provide their own.