Fix no-floating-promises lint violations
When our ESLint packages were updated, one of the rules that got enabled
was `no-floating-promises`, which is violated if a promise is not
properly awaited. To minimize alignment changes, this rule was
overridden at the time to produce a warning rather than an error.

This commit corrects code that violates `no-floating-promises` by either
awaiting promises where it is possible to do so, or by using a new
utility function, `logRejection`, to allow the promise to continue
unawaited while logging any possible rejections. This allows us to
remove the ESLint override.