Hide `merged-packages/` from ESLint/Prettier (#1524)
When we bring new libraries into this monorepo, we will first place them
in `merged-packages`. This is a special directory we are creating that
will be hidden from ESLint and Prettier (and also Yarn, since it's not
in the set of known workspaces). This allows us to clean up the packages
and get them passing the lint step before we properly move them to
`packages`.