Publish preview builds to NPM instead of GitHub (#1622)
Currently, when publishing core packages, we use NPM for "production"
releases but GitHub Package Registry for "preview" builds. This works to
keep the list of production versions free of clutter and mitigate the
risk of publishing preview builds as production releases accidentally,
but it also introduces a major pain point.

Usually in order to access a package with a preview build, the developer
must instruct Yarn to use GitHub Package Registry instead of NPM along
with an authentication token, and then update the version of the package
in the manifest to use the preview version. If, however, the package
introduces a new transitive dependency that is not available in preview
form, the developer must undo the changes to the Yarn configuration or
even temporarily add the dependency explicitly to the project in order
to fully install it. In some cases, these steps don't work at all.

To remove this pain point, this commit changes how preview builds are
published so that instead of using a separate registry, we use a
separate NPM scope. This preserves the separation between preview builds
and non-preview builds, and it means that the developer can mix
non-preview and preview versions in the manifest without needing to
reconfigure Yarn.

Co-authored-by: Mark Stacey <markjstacey@gmail.com>