Update preview builds instructions (#1629)
The existing instructions for using preview builds did not quite work.
To prevent errors we need to redirect two kinds of references to preview
builds: import statements and transitive dependencies. The best way to
achieve this is to make use of `resolutions`/`overrides` instead of
changing the dependency lines directly. This replaces the production
version of a package with its preview build across the entire dependency
tree.