Recommend focused resolutions for preview builds
When using a preview version, we don't want to replace the production
version across the dependency tree; we only really want to target its
direct usage by the project in question. This commit updates the
instructions for using preview builds to ensure that this happens by
asking the developer include a specific version on the left-hand side of
the resolution entry when updating `package.json`.