Rename transaction object to txParams in the persisted transaction metadata (#1651)
## Explanation
Rename the `transaction` object to `txParams` in the persisted
transaction metadata as part of the unification initiative between the
core and extension.

The following task will take place to handle the migration and adoption
of the patch/release in the mobile client.



## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/transaction-controller`

- **BREAKING**: rename the `transaction` object to `txParams` in the
persisted `TransactionMeta`.