Fix `addNewAccountForKeyring` for `CustodyKeyring` (#1694)
## Explanation
Currently, the `addNewAccountForKeyring` method calls
`keyring.getAccounts()` (where `keyring` is the arg passed to the func)
to store a list of accounts and later diff it with a new one to
determine the account added in the operation, and then return it.

### Problem
This strategy to find the new account does not work with
`CustodyKeyring`:
`CustodyKeyring.getAccounts()` returns a *reference* to its
`this.accounts` state property, which is stored in `oldAccounts`. The
reference is created
[here](https://github.com/consensys-vertical-apps/metamask-institutional/blob/15610d6d184e11a47cb8b2bbb0d93b9c12763172/packages/custodyKeyring/src/CustodyKeyring.ts#L172).
As this is a reference, when we call `CustodyKeyring.getAccounts()` the
second time and compare it with `oldAccounts`, what we are actually
doing is compare an object with itself (instead of _an old version of
itself_): this makes `addedAccountAddress` undefined, and the later
`assertIsStrictHexString` throws an error.

### Solution
The easy solution could be, as this PR suggests, to use
`this.getAccounts()` instead of `keyring.getAccounts()`, as the array
returned by KeyringController will be a copied object, and we will be
able to later call `this.getAccounts()` and compare the two copies, that
will present hopefully a difference of one address that will be assigned
to `addedAccountAddress`

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->

## References

<!--
Are there any issues that this pull request is tied to? Are there other
links that reviewers should consult to understand these changes better?

For example:

* Fixes #12345
* Related to #67890
-->
* Fixes #1700 

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/keyring-controller`

- **FIXED**: Fixed `addNewAccountForKeyring` for Keyrings like
`CustodyKeyring`, which returns referenced objects when calling
`getAccounts()`

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate