Subscribe to `QRKeyring` store and publish updates (#1702)
## Explanation
Currently, the extension is directly subscribing to the `QRKeyring`
store, to determine whether to open the modal for syncing and signing.

This raises concerns when the instance of the QRKeyring will be held by
`KeyringController` instead of the client because it will be tricky to
handle `subscribe` and `unsubscribe` in the keyring's lifecycle, risking
to have multiple subscriptions to the same QRKeyring instance or even no
subscriptions at all!

This PR gives this responsibility to `KeyringController`.
`KeyringController` will subscribe to the QRKeyring instance in these
methods:
- `submitPassword` - because here we unlock the vault and potentially
find a QRKeyring (even in case of re-construction from backup)
- `submitEncryptionKey` - same as above
- `getOrAddQRKeyring` - because here we explicitly create a new
QRKeyring if it doesn't exist
- `addNewKeyring(type)` - when `type` is of a QRKeyring then it will
return `getOrAddQRKeyring`, to avoid creating multiple keyring instances
(and multiple subscriptions)

`KeyringController` will unsubscribe from the `QRKeyring` instance (if
there's one ) when calling `setLocked()`.

Moreover, `KeyringController:qrKeyringStateChange` event is now
available on the messenger

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->

## References

<!--
Are there any issues that this pull request is tied to? Are there other
links that reviewers should consult to understand these changes better?

For example:

* Fixes #12345
* Related to #67890
-->

* Related to https://github.com/MetaMask/metamask-extension/issues/18776
* Related to https://github.com/MetaMask/metamask-extension/pull/20502

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/keyring-controller`

- **BREAKING**: `addNewKeyring(type)` return type changed from
`Promise<Keyring<Json>>` to `Promise<unknown>`
- When calling with QRKeyring `type` the keyring instance is retrieved
or created (no multiple QRKeyring instances possible)
- **ADDED**: Add `getQRKeyring(): QRKeyring | undefined` method
- **ADDED**: Add `KeyringController:qrKeyringStateChange` messenger
event
The event emits updates from the internal `QRKeyring` instance, if
there's one

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate