chore: Update `typedoc` and related packages (#1717)
## Explanation

The `typedoc` package has been updated to match the version used by the
module template. This was done in preparation for updating the
TypeScript version used by the core monorepo as well. The current
version of `typedoc` we use doesn't support TypeScript v4.8.

The `typedoc-plugin-missing-exports` package required an update to
maintain compatibility with `typedoc`.

In both cases, none of the breaking changes affect our usage of the
package. Mostly they include improvements to the generated
documentation, and dropping support for older Node.js and TypeScript
versions.

`typedoc` changelog:
https://github.com/TypeStrong/typedoc/releases/tag/v0.23.0
`typedoc-plugin-missing-exports` changelog:
https://github.com/Gerrit0/typedoc-plugin-missing-exports/blob/main/CHANGELOG.md#0230-2022-06-26

## References

None

## Changelog

N/A

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate