Added polling-controller (#1703)
## Explanation

> Originally was in https://github.com/MetaMask/core/pull/1673 but
pulled out to get this in on its own.

Adds an abstract class (currently named `PollingController`. The start
and stop methods are parameterized by `networkClientId`'s and
`pollingToken`'s and polling intervals are stored in class variables by
`chainId` so that multiple `networkClients`/`chainIds` can poll
simultaneously. `executePoll` is an abstract method to be implemented by
the controller itself so that this pattern could be generalized and be
agnostic to what is being executed on the polling interval.


## References
Related to https://github.com/MetaMask/core/pull/1673
Related to https://github.com/MetaMask/MetaMask-planning/issues/1314

---------

Co-authored-by: Alex Donesky <adonesky@gmail.com>