Add EIP-1559 compatibility validation for transaction creation (#1693)
## Explanation

This PR aims to include a validation existent in the Extension in the `TransactionController` core. The validation gets the current network and current account to check if they support EIP-1559 transactions. If the transaction is an EIP-1559 transaction, meaning it contains `maxFeePerGas` and `maxPriorityFeePerGas` values, but the network or account does not support EIP-1559, then an error is thrown.

In order to support it two callbacks are passed to the controller constructor `getCurrentAccountEIP1559Compatibility` and
`getCurrentNetworkEIP1559Compatibility`.

## Changelog

### `@metamask/transaction-controller`

- **BREAKING**: Add required `getCurrentAccountEIP1559Compatibility` and `getCurrentNetworkEIP1559Compatibility` callback arguments to constructor.