Migrate json rpc middleware stream repo (#1709)
## Explanation

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->
We would like to migrate most MetaMask libraries into core monorepo to
make them easier to maintain.

In this PR, we are migrating json-rpc-middleware-stream into core
monorepo. In the process of migration, we don't want to remove / loose
the original GIT history, but to preserve the history of the library as
though the entire project had been located in a
merged-packages/json-rpc-middleware-stream directory all along.

We chose to place json-rpc-middleware-stream initially in the
merged-packages and not packages so we can hide it from Yarn, ESLint,
Prettier, and TypeScript. This allows us to clean it up and conform it
to how we've configured this repo before we properly incorporate it.

## References

<!--
Are there any issues that this pull request is tied to? Are there other
links that reviewers should consult to understand these changes better?

For example:

* Fixes #12345
* Related to #67890
-->
Fixes
https://app.zenhub.com/workspaces/shared-libraries-621e46b4d7103800171d1b02/issues/gh/metamask/core/1680

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->


### `@metamask/json-rpc-middleware-stream`

- **Migration**: migrating to core monorepo

## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: kumavis <aaron@kumavis.me>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
Co-authored-by: Whymarrh Whitby <whymarrh.whitby@gmail.com>
Co-authored-by: Erik Marks <25517051+rekmarks@users.noreply.github.com>
Co-authored-by: Mark Stacey <markjstacey@gmail.com>
Co-authored-by: ryanml <ryanlanese@gmail.com>
Co-authored-by: Zbyszek Tenerowicz <naugtur@gmail.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Jyoti Puri <jyotipuri@gmail.com>
Co-authored-by: github-actions <github-actions@github.com>
Co-authored-by: Zach Pomerantz <zzmp@uniswap.org>
Co-authored-by: legobeat <109787230+legobeat@users.noreply.github.com>
Co-authored-by: legobt <6wbvkn0j@anonaddy.me>
Co-authored-by: Frederik Bolding <frederik.bolding@gmail.com>