Add validation to check origin permissions (#1722)
## Explanation
This PR aims to add validation in `addTransaction` to check the selected
address or origin permissions. Passed to the controller constructor
`getPermittedAccounts` callback to perform the validation when the
origin is not internal.

**Context**
The permitted accounts for an origin are determined using the
`getPermittedAccounts` callback passed to the controller constructor,
which in turn uses the `PermissionController`.
- If the origin is internal, does it match the selected account?
- If the origin is not internal, does it match the permitted accounts
for that origin?